Refactored server routes and controllers
Fixes #4
