+++ server/controllers/homeController.js
@@ -1,0 +1,7 @@
+module.exports = {
+
+  homepage: function(req, res, next) {
+    res.redirect('/homepage.html');
+  }
+
+};
